feat: 2nd try for multimodules
