+++ news-service/build.gradle
@@ -6,7 +6,6 @@ plugins {
 }
 
 group = 'ru.clevertec'
-version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '17'
 
 repositories {
@@ -22,4 +21,8 @@ dependencies {
 
 tasks.named('test') {
     useJUnitPlatform()
+}
+
+tasks.register('prepareKotlinBuildScriptModel') {
+
 }
