+++ news-service/src/main/java/NewsServiceApplication.java
@@ -1,5 +1,3 @@
-package ru.clevertec.newsservice;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
