+++ news-service/src/main/java/dao/CommentRepository.java
@@ -1,4 +1,4 @@
-package ru.clevertec.newsservice.dao;
+package dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
