+++ news-service/src/main/java/entity/Comment.java
@@ -1,11 +1,12 @@
-package ru.clevertec.newsservice.entity;
+package entity;
+
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,16 +14,14 @@ import lombok.Setter;
 import lombok.ToString;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@Entity
-@Table(name = "news")
 @ToString
-public class News implements BaseEntity<Long>{
+@Entity
+@Table(name = "comment")
+public class Comment implements BaseEntity<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,11 +30,12 @@ public class News implements BaseEntity<Long>{
     @Column(name = "creation_time")
     private LocalDateTime creationTime;
 
-    private String title;
-
     private String subject;
 
+    @ManyToOne
+    private User user;
+
+    @ManyToOne
     @ToString.Exclude
-    @OneToMany
-    private List<Comment> comments = new ArrayList<>();
+    private News news;
 }
