+++ news-service/src/main/java/entity/News.java
@@ -1,12 +1,11 @@
-package ru.clevertec.newsservice.entity;
-
+package entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,14 +13,16 @@ import lombok.Setter;
 import lombok.ToString;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@ToString
 @Entity
-@Table(name = "comment")
-public class Comment implements BaseEntity<Long> {
+@Table(name = "news")
+@ToString
+public class News implements BaseEntity<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -30,12 +31,11 @@ public class Comment implements BaseEntity<Long> {
     @Column(name = "creation_time")
     private LocalDateTime creationTime;
 
+    private String title;
+
     private String subject;
 
-    @ManyToOne
-    private User user;
-
-    @ManyToOne
     @ToString.Exclude
-    private News news;
+    @OneToMany
+    private List<Comment> comments = new ArrayList<>();
 }
