+++ news-service/src/main/java/entity/User.java
@@ -1,6 +1,5 @@
-package ru.clevertec.newsservice.entity;
+package entity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
