--- news-service/src/test/java/ru/clevertec/newsservice/NewsServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.clevertec.newsservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NewsServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
