+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'news'
-include('news-service')
+include 'news-service'
+
